Fixed scrollbar problem with enable()/disable().  They just weren't checking whether scrollbar was NULL or not.
Also changed activate() method to save the original text even if selection
was via mouse.  That seems to be what the original intent of the thing was.  Keeping track of a modified flag seems like it would be a much better way to check for whether a change was made, though.